+++ lib/libft/src/ft_vect_dot.c
@@ -3,18 +3,18 @@
 /*                                                        :::      ::::::::   */
 /*   ft_vect_dot.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: pabmart2 <pabmart2@student.42malaga.com    +#+  +:+       +#+        */
+/*   By: pablo <pablo@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/29 20:04:26 by pablo             #+#    #+#             */
-/*   Updated: 2025/01/30 16:13:14 by pabmart2         ###   ########.fr       */
+/*   Updated: 2025/02/07 13:33:53 by pablo            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../include/libft.h"
 
-double	ft_vect_dot(double *v1, double *v2, int d)
+double	ft_vect_dot(double *v1, double *v2, size_t d)
 {
-	int		i;
+	size_t		i;
 	double	result;
 
 	i = 0;
